+++ pyspaces/process.py
@@ -1,5 +1,11 @@
 #!/usr/bin/env python
 # coding=utf-8
+"""This is part of [pyspaces](https://github.com/Friz-zy/pyspaces)
+
+License: MIT or BSD or Apache 2.0
+Copyright (c) 2014 Filipp Kucheryavy aka Frizzy <filipp.s.frizzy@gmail.com>
+
+"""
 
 
 from os import chroot