+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changed
 - docstrings
 - error handling
+- add multilicensing: now pyspaces under MIT or BSD or Apache 2.0
 
 ## 1.2.3 - 2015-04-30
 ## Added